requirements.txt
python-dotenv==0.19.2
requests==2.26.0
dspy-ai==2.0.4
openai==0.27.0
